# services/platform/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tale Platform API",
    "version": "1.2.0",
...
  },
  "servers": [
    {
      "url": "",
      "description": "Same origin"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "Look up a project by external item id",
        "description": "A lookup door, not a list-all: `externalItemId` is REQUIRED (400 without it). Answers at most one project — the key is unique per organization. A project the key holder cannot see answers the same empty list as no match. The row carries `archivedAt` when the match is archived, so callers can detect an archived project holding the key. Shares the general REST bucket (120/min, keyed on the key holder).",
        "operationId": "lookupProjects",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "X-Organization-Slug",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "The organization to operate on. STRICTLY REQUIRED when the key holder belongs to more than one organization — without it such requests are refused with 400 (the last-active-org fallback is never used here). Single-organization keys may omit it."
          },
          {
            "name": "externalItemId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The caller-owned external key the project was created with"
          }
        ],
        "responses": {
          "200": {
            "description": "Zero or one matching project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["projects"],
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Project"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (malformed body or parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Retry-After names the wait)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create a project",
        "description": "Requires the org editor role. `externalItemId` is an opaque caller-owned key, stored trimmed, unique per organization regardless of lifecycle — a conflict against an archived project is still a 409. A blank or over-long value is a 400. `key` (the task-identifier prefix) is derived from the name when omitted; a derived key that collides gets a numeric suffix, and a name no key can be derived from creates the project keyless — only an EXPLICIT `key` conflict answers 409.",
        "operationId": "createProject",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "X-Organization-Slug",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "The organization to operate on. STRICTLY REQUIRED when the key holder belongs to more than one organization — without it such requests are refused with 400 (the last-active-org fallback is never used here). Single-organization keys may omit it."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 80
                  },
                  "externalItemId": {
                    "type": "string",
                    "maxLength": 256
                  },
                  "key": {
                    "type": "string",
                    "description": "Immutable 2-6 char task-key prefix. Omitted: derived from the name, suffixed on collision, keyless when underivable. Explicit: a collision is a 409."
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["project"],
                  "properties": {
                    "project": {
                      "$ref": "#/components/schemas/Project"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (malformed body or parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The key holder is not an org editor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate externalItemId or project key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Retry-After names the wait)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/automations": {
      "get": {
        "tags": ["Automations"],
        "summary": "List automations",
        "description": "Every automation of the organization, by name — a complete set, not paginated (an organization holds tens of automations, not thousands). `projectIds` is the binding set: empty means org-level.",
        "operationId": "listAutomations",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The organization’s automations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["automations"],
                  "properties": {
                    "automations": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AutomationSummary"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (malformed body or parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Retry-After names the wait)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/automations/{name}/versions": {
      "get": {
        "tags": ["Automations"],
        "summary": "List an automation’s versions",
        "description": "The immutable version history, complete — not paginated.",
        "operationId": "listAutomationVersions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The automation name — a `/`-separated slug written with `__` in place of each `/` (`billing/dunning` travels as `billing__dunning`). Responses always carry the real slug."
          }
        ],
        "responses": {
          "200": {
            "description": "Immutable version history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["versions", "name"],
                  "properties": {
                    "versions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AutomationVersion"
                      }
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (malformed body or parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Retry-After names the wait)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/automations/{name}/runs": {
      "get": {
        "tags": ["Automations"],
        "summary": "List an automation’s runs",
        "description": "The newest runs first — a bounded WINDOW (`limit` 1..200, default 50), not a cursor walk; each row is the full run. To follow one run, poll `GET /api/v1/runs/{runId}`.",
        "operationId": "listAutomationRuns",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The automation name — a `/`-separated slug written with `__` in place of each `/` (`billing/dunning` travels as `billing__dunning`). Responses always carry the real slug."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "How many of the newest runs to answer, 1..200 (default 50)"
          }
        ],
        "responses": {
          "200": {
            "description": "Newest runs first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["runs"],
                  "properties": {
                    "runs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Run"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (malformed body or parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Retry-After names the wait)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Automations"],
        "summary": "Start a run of the deployed version",
        "description": "Answers 202 with the run’s identity rather than its result: a run is durable and may take minutes, so poll `GET /api/v1/runs/{runId}`. Starting a run needs no trigger — the API key is the entitlement. `mode` defaults to `live`, which requires the key holder to have the developer capability; a `mock` run needs only membership. `projectId` scopes the run to a project (required, and restricted to the bound set, when the automation is bound to several). Charged against the execute bucket (20/min) on top of the general one.",
        "operationId": "startAutomationRun",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The automation name — a `/`-separated slug written with `__` in place of each `/` (`billing/dunning` travels as `billing__dunning`). Responses always carry the real slug."
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "input": {
                    "type": "object",
                    "description": "The run’s input value"
                  },
                  "mode": {
                    "type": "string",
                    "enum": ["live", "mock"],
                    "default": "live"
                  },
                  "version": {
                    "type": "integer",
                    "description": "Run a specific version (the deployed one when omitted)"
                  },
                  "projectId": {
                    "type": "string",
                    "description": "The project the run operates in"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Run started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["runId", "version", "name", "mode"],
                  "properties": {
                    "runId": {
                      "type": "string"
                    },
                    "version": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "string",
                      "enum": ["live", "mock"]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (malformed body or parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "A live run needs the developer capability",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "The automation has no deployed version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Retry-After names the wait)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/runs/{runId}": {
      "get": {
        "tags": ["Runs"],
        "summary": "Get one run in full",
        "operationId": "getRun",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The run ID a start call returned"
          }
        ],
        "responses": {
          "200": {
            "description": "The run: status, output, trace, effects, checkpoints",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (malformed body or parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Run not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Retry-After names the wait)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{runId}/cancel": {
      "post": {
        "tags": ["Runs"],
        "summary": "Stop a run at its next node boundary",
        "description": "Requires the developer capability. Work a node already performed is not undone.",
        "operationId": "cancelRun",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The run ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Cancellation result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["cancelled"],
                  "properties": {
                    "cancelled": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (malformed body or parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Needs the developer capability",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Run not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (Retry-After names the wait)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "An organization API key. Create one in Settings → API."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "What went wrong"
          },
          "code": {
            "type": "string",
            "description": "A stable refusal code, when the domain has one"
          }
        }
      },
...
      "Project": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "key": {
            "type": "string",
            "description": "Immutable task-identifier prefix"
          },
          "description": {
            "type": "string"
          },
          "externalItemId": {
            "type": "string",
            "description": "Caller-owned external key, unique per organization"
          },
          "archivedAt": {
            "type": "number",
            "description": "Present (epoch ms) when the project is archived — the archived marker for lookups"
          }
        }
      },
...
      "AutomationSummary": {
        "type": "object",
        "required": ["name", "latestVersion", "deployedVersion", "projectIds"],
        "properties": {
          "name": {
            "type": "string",
            "description": "The real `/`-slug"
          },
          "latestVersion": {
            "type": "integer"
          },
          "deployedVersion": {
            "type": "integer",
            "description": "null while nothing is deployed",
            "nullable": true
          },
          "presentation": {
            "description": "The newest version’s display block, if authored"
          },
          "projectIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The binding set; empty means org-level"
          }
        }
      },
...
      "AutomationVersion": {
        "type": "object",
        "required": ["version", "createdBy", "createdAt"],
        "properties": {
          "version": {
            "type": "integer"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "testsPassed": {
            "type": "boolean",
            "nullable": true
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "number",
            "description": "Epoch ms"
          }
        }
      },
...
    }
  }
}